server/requirements.txt
psycopg2-binary>=2.9.10
pgvector>=0.2.4
openai>=1.35.0
numpy>=2.1.0
pandas>=2.2.3
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.3
sqlalchemy>=2.0.25
asyncpg>=0.29.0